Sampling functions have been put into a library. The config() function is still a work in progress.
Config is designed so that at any point, "config" can be sent over UART and the multi-rate loop will halt. From there additional UART communication occurs so that the sampling parameters (window, rate) can be changed on the fly. After that the sampling parameters will be saved to an emulated EEPROM in flash and the C-CAN will reset itself. On boot the sampling parameters are read from the EEPROM; this way the C-CAN's can be *reconfigured* rather than reprogrammed to change how it functions - it would be tedious to reprogram 5+ boards individually. The final config function will work over CAN so that configuration can be initiated from a RPi that is connected to each C-CAN. That way the vehicle will have a "single point of entry" for configuration, ex. SSH into the RPi and configure.