USB bootloader added.
Now that there are some C-CANs going on the test mule, some of the configuration has been changed. Notably, the device is programmed via the PSoC Creator Bootloader Host tool. In addition, an external crystal oscillator will be used so that the CAN timing is more accurate (PSoC Creator warns that the internal oscillator is not accurate enough for high speed CAN).